move cancel prompt
